device/requirements.dev.txt
appdirs==1.4.4
black==21.5b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.3.1
grove.py==0.6
idna==2.10
mypy-extensions==0.4.3
numpy==1.20.2
paho-mqtt==1.5.1
pathspec==0.8.1
PyAudio==0.2.11
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.17.0
regex==2021.4.4
requests==2.25.1
scipy==1.6.3
six==1.15.0
smbus2==0.4.1
toml==0.10.2
urllib3==1.26.4
